+++ packages/utils/src/index.ts
@@ -6,4 +6,5 @@ export * from './internals/memory-info';
 export * from './internals/debug';
 export * as social from './internals/social';
 export * from './internals/typedefs';
+export * from './internals/error_tracker';
 export { Dictionary, Awaitable, Constructor } from '@crawlee/types';
